Week08/lab-solutions/Exercise-6/requirements.txt
Flask==0.10.1
python-slugify==0.0.9
